backend/requirements.txt
# AI Career Assistant - Backend Dependencies
# Python 3.11+ Required

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=12.0

# Database
sqlalchemy>=2.0.35
aiosqlite>=0.20.0

# AI & LLM
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.27
openai>=1.50.0

# Document Parsing
pdfplumber>=0.11.0
python-docx>=1.1.0

# Web Scraping
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Audio Processing
pydub>=0.25.0

# Configuration
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# Utilities
httpx>=0.27.0
aiofiles>=24.1.0
markdown>=3.7

# Testing (optional)
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
